requirements.txt
Django==4.1
djangorestframework==3.14.0
django-filter==23.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
daphne==4.0.0
psycopg2==2.9.9
# psycopg2-binary
drf-yasg==1.21.7
django-cors-headers==4.3.1
pillow==10.2.0

google-auth==2.26.1
requests

django-admin-interface==0.28.5
django-tenants==3.5.0
